Fixed cache problem
If the contents of the data direcory are changed, it appears that the cache is not updated because `cache=1` or `cache=2` don't work with `cache.extra` technique #994, see https://github.com/yihui/knitr/issues/994 and https://yihui.name/knitr/demo/cache/#more-granular-cache. This problem is solved by splitting the code chunck into two code chunks, one of which does nothing more that scanning the status of the data directory.